refactor : TableGroup 내 orderTables 필드 일급컬렉션 구현
